requirements.txt
Flask==2.1.2
requests==2.27.1
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.39.0
selenium==4.1.0
